move config to src folder
