+++ src/models/github/addLabels.js
@@ -1,7 +1,7 @@
 const octokit = require("@octokit/rest")()
 const {
   github: { token, owner, repo },
-} = require("../../config/config")
+} = require("../../config")
 
 module.exports = async function addLabel(issueNum, labels) {
   octokit.authenticate({
